Re-organize the versioned ui api code files.
No-Issue
