+++ galaxy_ng/app/api/ui/v1/serializers/search.py
