+++ galaxy_ng/app/api/ui/v1/urls.py
