+++ galaxy_ng/app/api/ui/v1/views/ai_index.py
