+++ galaxy_ng/app/api/ui/v1/views/landing_page.py
