+++ galaxy_ng/app/api/ui/v1/views/search.py
