+++ galaxy_ng/app/api/ui/v1/views/settings.py
