+++ galaxy_ng/app/api/ui/v2/views.py
